Do a pass of compiler warning fixes.
This is really making me question not using 64-bit integers more.

+++ Kernel/FileSystem/FileSystem.cpp
@@ -77,7 +77,7 @@ FS::DirectoryEntry::DirectoryEntry(const char* n, InodeIdentifier i, byte ft)
     name[name_length] = '\0';
 }
 
-FS::DirectoryEntry::DirectoryEntry(const char* n, size_t nl, InodeIdentifier i, byte ft)
+FS::DirectoryEntry::DirectoryEntry(const char* n, int nl, InodeIdentifier i, byte ft)
     : name_length(nl)
     , inode(i)
     , file_type(ft)
